+++ python/beeai_framework/adapters/beeai_platform/agents/types.py
@@ -2,10 +2,16 @@
 # SPDX-License-Identifier: Apache-2.0
 
 
-from acp_sdk.models.models import Event
+try:
+    import a2a.types as a2a_types
+except ModuleNotFoundError as e:
+    raise ModuleNotFoundError(
+        "Optional module [beeai-platform] not found.\nRun 'pip install \"beeai-framework[beeai-platform]\"' to install."
+    ) from e
+
 
 from beeai_framework.agents import AgentOutput
 
 
 class BeeAIPlatformAgentOutput(AgentOutput):
-    event: Event
+    event: a2a_types.SendStreamingMessageResponse | a2a_types.SendMessageResponse